Add files via upload
The above file is a web app that allows the user to input comma-separated list of Morning Line Odds for a horse race and then input at about 3 minutes to post, the Tote Board Odds and the users bankroll.  The script then computes the best pair of horses for an Exacta Box Ticket and using the Kelly Criterion suggests a wager or wagers based upon the Bayesian statistics of the input horses and the users bankroll.  This app is purely for entertainment and educational use only, Gambling involves risk of complete loss, the author of this code gives no warranty expressed or implied, nor any guarantee expressed or implied as to performance, function, utility, profitability.  It was purely a project intended to teach me how to code in a way I could have fun with.